Organized code and test directories
