+++ code/old/BrowsingManager.py
