+++ code/old/PDFMakerForTablet.py
@@ -63,7 +63,7 @@ class PDFMakerForTablet (PDFMaker):
 
 
 def test():
-    UG_urls_file_path = '../input_urls/TAB4U-test.txt'
+    UG_urls_file_path = '../../input_urls/TAB4U-test.txt'
     tablet_chord_writer = PDFMakerForTablet(UG_urls_file_path, 'TAB4U')
     pdf_file_path = 'output_pdfs/'+UG_urls_file_path.split('/')[-1].split('.')[0]+'.pdf'
     tablet_chord_writer.make_pdf(pdf_file_path)
